requirements.txt
flask>=3.1.1
python-dotenv>=1.1.0
gunicorn